fix routing order to booking
