+++ app/database/mysql.go
@@ -7,7 +7,6 @@ import (
 	bd "KosKita/features/booking/data"
 	md "KosKita/features/chat/data"
 	kd "KosKita/features/kos/data"
-	od "KosKita/features/order/data"
 	ud "KosKita/features/user/data"
 
 	"gorm.io/driver/mysql"
@@ -32,7 +31,6 @@ func InitDBMysql(cfg *config.AppConfig) *gorm.DB {
 		&kd.KosFacility{},
 		&kd.KosRule{},
 		&md.Chat{},
-		&od.Order{},
 	)
 
 	return DB
