+++ features/booking/handler/handler.go
@@ -5,90 +5,91 @@ import (
 	"KosKita/utils/middlewares"
 	"KosKita/utils/responses"
 	"fmt"
+	"log"
 	"net/http"
 
 	"github.com/labstack/echo/v4"
 )
 
-type BookHandler struct {
-	bookService booking.BookServiceInterface
+// _midtransController "BE-REPO-20/features/midtrans/controller"
+
+type BookingHandler struct {
+	bookingService booking.BookingServiceInterface
 }
 
-func New(bs booking.BookServiceInterface) *BookHandler {
-	return &BookHandler{
-		bookService: bs,
+func NewBooking(service booking.BookingServiceInterface) *BookingHandler {
+	return &BookingHandler{
+		bookingService: service,
 	}
 }
 
-func (handler *BookHandler) CreateBook(c echo.Context) error {
-	userIdLogin := middlewares.ExtractTokenUserId(c)
-	if userIdLogin == 0 {
-		return c.JSON(http.StatusUnauthorized, responses.WebResponse("Unauthorized user", nil))
+func (handler *BookingHandler) CreateBooking(c echo.Context) error {
+	idJWT := middlewares.ExtractTokenUserId(c)
+	if idJWT == 0 {
+		return c.JSON(http.StatusBadRequest, responses.WebResponse("unauthorized or jwt expired", nil))
 	}
 
-	newBook := BookRequest{}
-	errBind := c.Bind(&newBook)
+	newBooking := BookingRequest{}
+	errBind := c.Bind(&newBooking)
 	if errBind != nil {
 		return c.JSON(http.StatusBadRequest, responses.WebResponse("error bind data. data booking not valid", nil))
 	}
 
-	bookCore := RequestToCoreBook(newBook, uint(userIdLogin))
-	payment, errInsert := handler.bookService.Create(userIdLogin, bookCore)
+	bookingCore := RequestToCoreBooking(newBooking)
+	payment, errInsert := handler.bookingService.PostBooking(uint(idJWT), bookingCore)
 	if errInsert != nil {
-		return c.JSON(http.StatusInternalServerError, responses.WebResponse(errInsert.Error(), nil))
+		return c.JSON(http.StatusInternalServerError, responses.WebResponse("error insert booking "+errInsert.Error(), nil))
 	}
 
-	// result := CoreToResponseBook(payment)
-	result := BookingResponse{}
-	if payment != nil {
-		result = CoreToResponseBook(payment)
-	}
+	results := CoreToResponse(*payment)
 
-	return c.JSON(http.StatusOK, responses.WebResponse("success booking kos", result))
+	return c.JSON(http.StatusOK, responses.WebResponse("Success get booking.", results))
 }
 
-func (handler *BookHandler) CancelBooking(c echo.Context) error {
+func (handler *BookingHandler) GetBookings(c echo.Context) error {
+	idJWT := middlewares.ExtractTokenUserId(c)
+	if idJWT == 0 {
+		return c.JSON(http.StatusBadRequest, responses.WebResponse("unauthorized or jwt expired", nil))
+	}
+
+	results, err := handler.bookingService.GetBookings(uint(idJWT))
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, responses.WebResponse("Error booking. "+err.Error(), nil))
+	}
+
+	var bookingResults []BookingHistoryResponse
+	for _, result := range results {
+		bookingResults = append(bookingResults, CoreToResponseBookingHistory(&result))
+	}
+
+	return c.JSON(http.StatusOK, responses.WebResponse("Success get booking.", bookingResults))
+}
+
+func (handler *BookingHandler) CancelBookingById(c echo.Context) error {
 	userIdLogin := middlewares.ExtractTokenUserId(c)
 	if userIdLogin == 0 {
 		return c.JSON(http.StatusUnauthorized, responses.WebResponse("Unauthorized user", nil))
 	}
 
-	bookingId := c.Param("id")
-	fmt.Println(bookingId)
+	bookingId := c.Param("booking_id")
 
 	updateBookingStatus := CancelBookingRequest{}
 	errBind := c.Bind(&updateBookingStatus)
 	if errBind != nil {
 		return c.JSON(http.StatusBadRequest, responses.WebResponse("error bind data. data not valid", nil))
 	}
-
+	fmt.Println("Booking id::", bookingId)
 	bookingCore := CancelRequestToCoreBooking(updateBookingStatus)
-	errCancel := handler.bookService.CancelBooking(userIdLogin, bookingId, bookingCore)
+	errCancel := handler.bookingService.CancelBooking(userIdLogin, bookingId, bookingCore)
 	if errCancel != nil {
-		return c.JSON(http.StatusInternalServerError, responses.WebResponse("error cancel booking "+errCancel.Error(), nil))
+		return c.JSON(http.StatusBadRequest, responses.WebResponse("error cancel booking "+errCancel.Error(), nil))
 	}
 
 	return c.JSON(http.StatusOK, responses.WebResponse("success cancel booking", nil))
 }
 
-func (handler *BookHandler) GetBooking(c echo.Context) error {
-	idJWT := middlewares.ExtractTokenUserId(c)
-	if idJWT == 0 {
-		return c.JSON(http.StatusBadRequest, responses.WebResponse("unauthorized or jwt expired", nil))
-	}
+func (handler *BookingHandler) WebhoocksNotification(c echo.Context) error {
 
-	results, err := handler.bookService.GetBooking(uint(idJWT))
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, responses.WebResponse("Error booking. "+err.Error(), nil))
-	}
-	var bookingResults []BookingHistoryResponse
-	for _, result := range results {
-		bookingResults = append(bookingResults, CoreToResponseBookHistory(&result))
-	}
-	return c.JSON(http.StatusOK, responses.WebResponse("Success get booking.", bookingResults))
-}
-
-func (handler *BookHandler) WebhoocksNotification(c echo.Context) error {
 	var webhoocksReq = WebhoocksRequest{}
 	errBind := c.Bind(&webhoocksReq)
 	if errBind != nil {
@@ -96,10 +97,11 @@ func (handler *BookHandler) WebhoocksNotification(c echo.Context) error {
 	}
 
 	bookingCore := WebhoocksRequestToCore(webhoocksReq)
-	err := handler.bookService.WebhoocksData(bookingCore)
+	err := handler.bookingService.WebhoocksService(bookingCore)
 	if err != nil {
-		return c.JSON(http.StatusInternalServerError, responses.WebResponse(err.Error(), nil))
+		return c.JSON(http.StatusInternalServerError, responses.WebResponse("error Notif "+err.Error(), nil))
 	}
 
-	return c.JSON(http.StatusOK, responses.WebResponse("success update status booking", nil))
+	log.Println("transaction success")
+	return c.JSON(http.StatusOK, responses.WebResponse("transaction success", nil))
 }
