+++ features/booking/handler/request.go
@@ -6,22 +6,35 @@ import (
 	"github.com/google/uuid"
 )
 
-type BookRequest struct {
+type BookingRequest struct {
+	ID              string
 	BoardingHouseId uint   `json:"kos_id"`
 	Method          string `json:"payment_type" form:"payment_type"`
 	Bank            string `json:"bank" form:"bank"`
+	StartDate       string `json:"start_date" form:"start_date"`
+}
+
+type CancelOrderRequest struct {
+	Status string `json:"status"`
 }
 
 type CancelBookingRequest struct {
 	Status string `json:"status"`
 }
 
-func RequestToCoreBook(input BookRequest, userIdLogin uint) booking.BookingCore {
+type WebhoocksRequest struct {
+	OrderID           string `json:"order_id"`
+	TransactionStatus string `json:"transaction_status"`
+	SignatureKey      string `json:"signature_key"`
+	TransactionTime   string `json:"transaction_time"`
+}
+
+func RequestToCoreBooking(input BookingRequest) booking.BookingCore {
 	return booking.BookingCore{
-		Code:            uuid.New().String(),
-		UserId:          userIdLogin,
+		ID:              uuid.New().String(),
 		BoardingHouseId: input.BoardingHouseId,
-		Method:          input.Method,
+		StartDate:       input.StartDate,
+		PaymentType:     input.Method,
 		Bank:            input.Bank,
 	}
 }
@@ -32,14 +45,10 @@ func CancelRequestToCoreBooking(input CancelBookingRequest) booking.BookingCore 
 	}
 }
 
-type WebhoocksRequest struct {
-	Code   string `json:"order_id"`
-	Status string `json:"transaction_status"`
-}
-
 func WebhoocksRequestToCore(input WebhoocksRequest) booking.BookingCore {
 	return booking.BookingCore{
-		Code:   input.Code,
-		Status: input.Status,
+		ID:     input.OrderID,
+		Status: input.TransactionStatus,
+		PaidAt: input.TransactionTime,
 	}
 }
