+++ utils/externalapi/midtrans.go
@@ -3,9 +3,9 @@ package externalapi
 import (
 	"KosKita/app/config"
 	"KosKita/features/booking"
-	"KosKita/features/order"
+
+	// "KosKita/features/order"
 	"errors"
-	"fmt"
 
 	mid "github.com/midtrans/midtrans-go"
 
@@ -15,8 +15,8 @@ import (
 type MidtransInterface interface {
 	NewOrderPayment(book booking.BookingCore) (*booking.BookingCore, error)
 	CancelOrderPayment(bookingId string) error
-	NewOrderPaymentOrder(order order.OrderCore) (*order.OrderCore, error)
-	CancelOrderPaymentOrder(orderId string) error
+	// NewOrderPaymentOrder(order order.OrderCore) (*order.OrderCore, error)
+	// CancelOrderPaymentOrder(orderId string) error
 }
 
 type midtrans struct {
@@ -38,7 +38,7 @@ func New() MidtransInterface {
 func (pay *midtrans) NewOrderPayment(book booking.BookingCore) (*booking.BookingCore, error) {
 	req := new(coreapi.ChargeReq)
 	req.TransactionDetails = mid.TransactionDetails{
-		OrderID:  book.Code,
+		OrderID:  book.ID,
 		GrossAmt: int64(book.Total),
 	}
 
@@ -84,7 +84,7 @@ func (pay *midtrans) NewOrderPayment(book booking.BookingCore) (*booking.Booking
 	}
 
 	if res.PaymentType != "" {
-		book.Method = res.PaymentType
+		book.PaymentType = res.PaymentType
 	}
 
 	if res.TransactionStatus != "" {
@@ -95,13 +95,9 @@ func (pay *midtrans) NewOrderPayment(book booking.BookingCore) (*booking.Booking
 		book.Status = res.TransactionStatus
 	}
 
-	// if expiredAt, err := time.Parse("2006-01-02 15:04:05", res.ExpiryTime); err != nil {
-	// 	return nil, err
-	// } else {
-	// 	book.ExpiredAt = expiredAt
-	// }
-
-	book.BookingTotal = book.Total
+	if res.ExpiryTime != "" {
+		book.ExpiredAt = res.ExpiryTime
+	}
 
 	return &book, nil
 }
@@ -111,86 +107,6 @@ func (pay *midtrans) CancelOrderPayment(bookingId string) error {
 	if res.StatusCode != "200" && res.StatusCode != "412" {
 		return errors.New(res.StatusMessage)
 	}
-
-	return nil
-}
-
-// NewOrderPayment implements Midtrans.
-func (pay *midtrans) NewOrderPaymentOrder(order order.OrderCore) (*order.OrderCore, error) {
-	req := new(coreapi.ChargeReq)
-	req.TransactionDetails = mid.TransactionDetails{
-		OrderID:  order.ID,
-		GrossAmt: int64(order.Total),
-	}
-
-	switch order.Bank {
-	case "bca":
-		req.PaymentType = coreapi.PaymentTypeBankTransfer
-		req.BankTransfer = &coreapi.BankTransferDetails{
-			Bank: mid.BankBca,
-		}
-	case "bni":
-		req.PaymentType = coreapi.PaymentTypeBankTransfer
-		req.BankTransfer = &coreapi.BankTransferDetails{
-			Bank: mid.BankBni,
-		}
-	case "bri":
-		req.PaymentType = coreapi.PaymentTypeBankTransfer
-		req.BankTransfer = &coreapi.BankTransferDetails{
-			Bank: mid.BankBri,
-		}
-	case "permata":
-		req.PaymentType = coreapi.PaymentTypeBankTransfer
-		req.BankTransfer = &coreapi.BankTransferDetails{
-			Bank: mid.BankPermata,
-		}
-	default:
-		return nil, errors.New("unsupported payment")
-	}
-
-	res, err := pay.client.ChargeTransaction(req)
-	if err != nil {
-		return nil, err
-	}
-	if res.StatusCode != "201" {
-		return nil, errors.New(res.StatusMessage)
-	}
-
-	if len(res.VaNumbers) == 1 {
-		order.VirtualNumber = res.VaNumbers[0].VANumber
-	}
-
-	if res.PermataVaNumber != "" {
-		order.VirtualNumber = res.PermataVaNumber
-	}
-
-	if res.PaymentType != "" {
-		order.PaymentType = res.PaymentType
-	}
-
-	if res.TransactionStatus != "" {
-		order.Status = res.TransactionStatus
-	}
-
-	// if expiredAt, err := time.Parse("2006-01-02 15:04:05", res.ExpiryTime); err != nil {
-	// 	return nil, err
-	// } else {
-	// 	order.ExpiredAt = expiredAt
-	// }
-	fmt.Println("expired time from Midtrans", res.ExpiryTime)
-	if res.ExpiryTime != "" {
-		order.ExpiredAt = res.ExpiryTime
-	}
-
-	return &order, nil
-}
-
-func (pay *midtrans) CancelOrderPaymentOrder(orderId string) error {
-	res, _ := pay.client.CancelTransaction(orderId)
-	if res.StatusCode != "200" && res.StatusCode != "412" {
-		return errors.New(res.StatusMessage)
-	}
-	fmt.Println("transaction time from Midtrans", res.TransactionTime)
 
 	return nil
 }
